refactor-register-shelter: cambio files y folders y adapto su uso
