--- src/RegisterShelter/components/index.js
@@ -1,4 +1,0 @@
-export { RegisterShelterSelect } from "./RegisterShelterSelect";
-export { RegisterShelterForm } from "./RegisterShelterForm";
-export { EyeIcon } from "./EyeIcon";
-export { RegisterShelterInput } from "./RegisterShelterInput";
