--- src/RegisterShelter/hooks/index.js
@@ -1,1 +1,0 @@
-export { useAddress } from "./useAddress";
