--- src/RegisterShelter/index.js
@@ -1,2 +1,0 @@
-export { RegisterShelterInput, EyeIcon ,RegisterShelterForm,RegisterShelterSelect} from "./components";
-export * from './data'
