--- src/RegisterShelter/style/RegisterShelterInput.module.css
@@ -1,61 +1,0 @@
-.registerShelterInput {
-  color: var(--brand-primary-01);
-  font-weight: 400;
-  outline: none;
-  border: none;
-  border-radius: 10px;
-  background-color: var(--background-lightmode-secondary);
-}
-.registerShelterTextarea {
-  color: var(--brand-primary-01);
-  font-weight: 400;
-  outline: none;
-  border: none;
-  border-radius: 10px;
-  background-color: var(--background-lightmode-secondary);
-}
-
-.registerShelterTextarea::placeholder {
-  color: var(--brand-neutro-01);
-}
-.registerShelterInput::placeholder {
-  color: var(--brand-neutro-01);
-}
-
-.registerShelterSpan {
-  width: 100%;
-  color: var(--system-error);
-  padding-left: 16px;
-  display: inline-block;
-}
-
-.registerShelterEyeButton {
-  background: none;
-  border: none;
-  padding: 0;
-  margin: 0;
-  cursor: pointer;
-  position: absolute;
-  transform: translateY(-50%);
-  top: 50%;
-  right: 0;
-  margin-right: 16px;
-}
-
-.registerShelterEyeButton:focus {
-  outline: none;
-}
-
-.registerShelterSvg {
-  width: 24px;
-  height: 24px;
-}
-
-.registerShelterError .registerShelterInput,
-.registerShelterError .registerShelterTextarea {
-  color: var(--system-error);
-  border: 1px solid var(--system-error);
-}
-.registerShelterError .registerShelterSvg {
-  stroke: var(--system-error);
-}
