+++ src/components/elements/index.js
@@ -1,10 +1,11 @@
-export { default as Select } from "react-select/base";
 export { default as ArrowBack } from "./ArrowBack";
 export { default as Check } from "./Check";
+export { default as Input } from "./Input";
 export { default as MyDatePicker } from "./DatePicker";
+export { default as Select } from "react-select/base";
+export { default as SelectGeneral } from "./Select";
 export { EyeIcon } from "./EyeIcon";
 export { HeaderButton } from "./HeaderButton";
-export { default as Input } from "./Input";
 export { MapPin } from "./MapPin";
 export { Pill } from "./Pill";
 export { Tag } from "./Tag";
