+++ src/hooks/index.js
@@ -1,2 +1,3 @@
+export { useFetchAddress } from "./useFetchAddress";
 export { useFetchPetByIdForShelter } from "./useFetchPetByIdForShelter";
 export { useFetchPetById } from "./useFetchPetById";
